backend/requirements-simple.txt
fastapi==0.110.1
uvicorn==0.25.0
python-dotenv==1.1.1
pydantic==2.12.0
razorpay==2.0.0
httpx==0.28.1
python-multipart==0.0.20
